livestream-scraper/requirements.txt
playwright>=1.40.0
pyyaml>=6.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
sqlalchemy>=2.0.0
tabulate>=0.9.0
